% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctrs}
\alias{ctrs}
\title{Names of the 11 countries included in the multiple-group analyses.}
\format{An object of class \code{character} of length 11.}
\usage{
data("ctrs")
}
\description{
A character vector containing the names.
}
\keyword{datasets}
