% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{id_subset}
\alias{id_subset}
\title{RowIDs of 500 randomly selected respondents per country.}
\format{
An object of class \code{integer} of length 18182.
}
\usage{
data("id_subset")
}
\description{
A random subsample of at most 500 respondents (with complete cases on the variables v45--v47) was selected in each country.
The vector \code{id_subset} contains the rowIDs of these respondents pertaining to \code{issp15$rowid}.
}
\examples{
subset(issp15, subset = rowid \%in\% id_subset)
}
\keyword{datasets}
