% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\title{The generated data and parameters of the simulation study.}
\format{An object of class \code{list} of length 7.}
\usage{
data("sim_data")
}
\description{
A list containing the data \code{sim_data$dat} and the data-generating parameters.
}
\keyword{datasets}
