% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_res}
\alias{sim_res}
\title{The Mplus output for the three models fitted in the simulation study.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 2 columns.}
\usage{
data("sim_res")
}
\description{
A tibble, where each of three rows contains the Mplus output in the column \code{output}.
}
\keyword{datasets}
