% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{sim_tree_data}
\alias{sim_tree_data}
\title{Generate Data for MCN Model}
\usage{
sim_tree_data(n = 200, items = 10, theta_vcov = NULL, betas = NULL,
  thetas = NULL, empirical = FALSE)
}
\arguments{
\item{n}{Numeric. Number of participants.}

\item{items}{Numeric. Number of items.}

\item{theta_vcov}{Covariance matrix for MRS, ERS, and target trait.}

\item{betas}{Jx3 matrix with item parameters beta pertaining to the three
dimensions (MRS, ERS, target trait).}
}
\value{
A list containing the generated matrix of responses \code{dat}, the
  generated person parameters \code{theta}, and the calculated response
  probabilities \code{prob} among other things.
}
\description{
Generate Data for MCN Model
}
\examples{
# generate data
n <- 20; i <- 10
betas <- cbind(runif(i, 0, 1), runif(i, 0, 1), rnorm(i, 0, 2))
gen <- sim_tree_data(n = n, items = i, betas = betas, theta_vcov = diag(3))
# barplot(table(factor(gen$dat[, 1], 0:4)))
}
